Refactor project structure: reorganize utils into common and presentation-focused modules
