+++ utils/common/gemini_service.py
