+++ utils/presentation/image_search_service.py
